chore(lint): Cleaning up type assertions and adding return types

+++ module/pack/bin/lib/assemble.ts
@@ -9,14 +9,14 @@ import { PackUtil } from './util';
 const MODULE_DIRS = ['src', 'bin', 'support', 'resources', 'index.ts', 'package.json', 'tsconfig.trv.json'];
 
 /**
- * Utils for assmbling
+ * Utils for assembling
  */
 export class AssembleUtil {
 
   /**
    * Minimize cached source files, by removing source mapping info
    */
-  static async cleanCache(cache: string) {
+  static async cleanCache(cache: string): Promise<void> {
     for (const el of await fs.readdir(cache)) {
       if (el.endsWith('.ts')) {
         const content = (await fs.readFile(`${cache}/${el}`, 'utf8')).replace(/\/\/# sourceMap.*/g, '');
@@ -28,10 +28,10 @@ export class AssembleUtil {
   /**
    * Truncate all app source files, and framework source files
    */
-  static async purgeSource(folders: string[]) {
+  static async purgeSource(folders: string[]): Promise<void> {
     for (const sub of folders) {
       for (const f of await ScanFs.scanDir({ testFile: x => x.endsWith('.ts'), testDir: x => true }, sub)) {
-        if (f.stats.isFile() && !f.module.startsWith('cli/')) {
+        if (f.stats?.isFile() && !f.module.startsWith('cli/')) {
           await fs.writeFile(f.file, '');
         }
       }
@@ -41,17 +41,17 @@ export class AssembleUtil {
   /**
    * Copy a module
    */
-  static async copyModule(root: string, target: string) {
+  static async copyModule(root: string, target: string): Promise<void> {
     for (const f of MODULE_DIRS) {
-      const stgt = PathUtil.resolveUnix(root, f);
-      const ftgt = PathUtil.resolveUnix(target, f);
-      const found = await FsUtil.exists(stgt);
+      const sourceFile = PathUtil.resolveUnix(root, f);
+      const targetFile = PathUtil.resolveUnix(target, f);
+      const found = await FsUtil.exists(sourceFile);
       if (found) {
         if (found.isFile()) {
-          await fs.copyFile(stgt, ftgt);
+          await fs.copyFile(sourceFile, targetFile);
         } else {
-          await fs.mkdir(ftgt, { recursive: true });
-          await FsUtil.copyRecursive(`${stgt}/*`, ftgt);
+          await fs.mkdir(targetFile, { recursive: true });
+          await FsUtil.copyRecursive(`${sourceFile}/*`, targetFile);
         }
       }
     }
@@ -60,10 +60,10 @@ export class AssembleUtil {
   /**
    * Purge workspace using file rules
    */
-  static async excludeFiles(root: string, files: string[]) {
+  static async excludeFiles(root: string, files: string[]): Promise<void> {
     const checker = PackUtil.excludeChecker(files, root);
     for (const el of await ScanFs.scanDir({ testDir: x => true, testFile: checker, withHidden: true }, root)) {
-      if (!el.stats.isFile()) { continue; }
+      if (!el.stats || !el.stats.isFile()) { continue; }
       try {
         await fs.unlink(el.file);
       } catch { }
@@ -71,9 +71,9 @@ export class AssembleUtil {
   }
 
   /**
-   * Copy over all prod dependnecies
+   * Copy over all prod dependencies
    */
-  static async copyDependencies(workspace: string, types: DepType[] = ['prod', 'opt', 'optPeer']) {
+  static async copyDependencies(workspace: string, types: DepType[] = ['prod', 'opt', 'optPeer']): Promise<void> {
 
     for (const el of await DependenciesUtil.resolveDependencies({ types })) {
       const sub = PathUtil.normalizeFrameworkPath(el.file, 'node_modules/')
@@ -103,7 +103,7 @@ export class AssembleUtil {
   /**
    * Compile workspace
    */
-  static async buildWorkspace(root: string, cacheDir: string) {
+  static async buildWorkspace(root: string, cacheDir: string): Promise<void> {
     await ExecUtil.spawn('node', ['./node_modules/@travetto/cli/bin/trv.js', 'build'],
       {
         cwd: root, isolatedEnv: true,
@@ -115,7 +115,7 @@ export class AssembleUtil {
   /**
    * Copy over added content
    */
-  static async copyAddedContent(workspace: string, files: Record<string, string>[]) {
+  static async copyAddedContent(workspace: string, files: Record<string, string>[]): Promise<void> {
     for (const a of files) {
       let [src, dest] = Object.entries(a)[0];
       [src, dest] = [PathUtil.resolveUnix(src), PathUtil.resolveUnix(workspace, dest)];
